# specs/001-warehouse-management/contracts/openapi.yaml
openapi: 3.1.2
...
info:
  title: Warehouse Manager API
  version: 1.0.0
  description: >-
    Authoritative HTTP contract for warehouse, route, sale, reconciliation, reporting, document, user, and
    printer-profile operations. Money and quantities are decimal strings. All protected authorization and business rules
    are enforced by the API.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /users:
    get:
      tags: [ Users ]
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Users visible to an administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '403':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: List users
    post:
      tags: [ Users ]
      operationId: createUser
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Create user
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [ Users ]
      operationId: getUser
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Get user
    patch:
      tags: [ Users ]
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated, activated, or deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Update user
...
  /printer-profiles:
    get:
      tags: [ Printers ]
      operationId: listPrinterProfiles
      responses:
        '200':
          description: Approved printer profiles available to the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrinterProfileListResponse'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: List printer profiles
    post:
      tags: [ Printers ]
      operationId: createPrinterProfile
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrinterProfileWriteRequest'
      responses:
        '201':
          description: Approved printer profile created by an administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrinterProfileResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Create printer profile
  /me/printer-preference:
    get:
      tags: [ Printers ]
      operationId: getMyPrinterPreference
      responses:
        '200':
          description: Current user's limited printer preference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrinterPreferenceResponse'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Get my printer preference
    put:
      tags: [ Printers ]
      operationId: setMyPrinterPreference
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrinterPreferenceRequest'
      responses:
        '200':
          description: Preference saved; physical permission remains browser-local
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrinterPreferenceResponse'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Set my printer preference
...
  /printer-profiles/{printerProfileId}:
    parameters:
      - name: printerProfileId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    patch:
      tags: [ Printers ]
      operationId: updatePrinterProfile
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrinterProfileUpdateRequest'
      responses:
        '200':
          description: Supported printer profile updated or deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrinterProfileResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Update printer profile
...
  /settings/business:
    get:
      tags: [ Catalog ]
      operationId: getBusinessSettings
      responses:
        '200':
          description: Business currency, timezone, partner rate, and rounding settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessSettingResponse'
        '403':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Get business settings
    patch:
      tags: [ Catalog ]
      operationId: updateBusinessSettings
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessSettingUpdateRequest'
      responses:
        '200':
          description: Settings updated for future operations; history remains unchanged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessSettingResponse'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
      summary: Update business settings
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-wm_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Per-session synchronizer token for an authenticated unsafe request
      schema:
        type: string
        minLength: 32
        maxLength: 512
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Search:
      name: search
      in: query
      schema:
        type: string
        maxLength: 120
    UserId:
      name: userId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  responses:
    Problem:
      description: RFC 9457 error; exact status and stable code identify the failure
      headers:
        X-Request-Id:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Uuid:
      type: string
      format: uuid
...
    UserRole:
      type: string
      enum: [ ADMINISTRATOR, DRIVER ]
    Problem:
      type: object
      additionalProperties: false
      required: [ type, title, status, code, instance ]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]+$'
        instance:
          type: string
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [ path, code, message ]
            properties:
              path:
                type: string
              code:
                type: string
              message:
                type: string
        requestId:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    PageInfo:
      type: object
      additionalProperties: false
      required: [ hasNextPage ]
      properties:
        hasNextPage:
          type: boolean
        nextCursor:
          type: [ string, 'null' ]
...
    User:
      type: object
      additionalProperties: false
      required: [ id, username, displayName, role, active, version, createdAt, updatedAt ]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        username:
          type: string
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        active:
          type: boolean
        version:
          type: integer
          minimum: 1
        archivedAt:
          type: [ string, 'null' ]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserCreateRequest:
      type: object
      additionalProperties: false
      required: [ username, displayName, role, password ]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 120
        displayName:
          type: string
          minLength: 1
          maxLength: 160
        role:
          $ref: '#/components/schemas/UserRole'
        password:
          type: string
          minLength: 12
          maxLength: 1024
    UserUpdateRequest:
      type: object
      additionalProperties: false
      required: [ expectedVersion ]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        displayName:
          type: string
          minLength: 1
          maxLength: 160
        role:
          $ref: '#/components/schemas/UserRole'
        active:
          type: boolean
        password:
          type: string
          minLength: 12
          maxLength: 1024
        reason:
          type: string
          minLength: 1
          maxLength: 500
    UserResponse:
      type: object
      additionalProperties: false
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/User'
    UserListResponse:
      type: object
      additionalProperties: false
      required: [ data, page ]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        page:
          $ref: '#/components/schemas/PageInfo'
...
    PrinterProfile:
      type: object
      additionalProperties: false
      required:
        [
          id,
          name,
          model,
          transport,
          serviceUuid,
          writeCharacteristicUuid,
          paperWidthMm,
          encoding,
          maxChunkBytes,
          active,
          version
        ]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
        model:
          type: string
        transport:
          const: WEB_BLUETOOTH_BLE
        serviceUuid:
          type: string
        writeCharacteristicUuid:
          type: string
        writeMode:
          type: string
          enum: [ WITH_RESPONSE, WITHOUT_RESPONSE ]
        commandDialect:
          type: string
        paperWidthMm:
          type: integer
          enum: [ 58, 80 ]
        encoding:
          type: string
        maxChunkBytes:
          type: integer
          minimum: 1
          maximum: 1024
        interChunkDelayMs:
          type: integer
          minimum: 0
          maximum: 5000
        active:
          type: boolean
        version:
          type: integer
          minimum: 1
    PrinterProfileWriteRequest:
      type: object
      additionalProperties: false
      required:
        [
          name,
          model,
          serviceUuid,
          writeCharacteristicUuid,
          writeMode,
          commandDialect,
          paperWidthMm,
          encoding,
          maxChunkBytes,
          interChunkDelayMs
        ]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        model:
          type: string
          minLength: 1
          maxLength: 120
        serviceUuid:
          type: string
          maxLength: 64
        writeCharacteristicUuid:
          type: string
          maxLength: 64
        writeMode:
          type: string
          enum: [ WITH_RESPONSE, WITHOUT_RESPONSE ]
        commandDialect:
          type: string
          maxLength: 80
        paperWidthMm:
          type: integer
          enum: [ 58, 80 ]
        encoding:
          type: string
          maxLength: 40
        maxChunkBytes:
          type: integer
          minimum: 1
          maximum: 1024
        interChunkDelayMs:
          type: integer
          minimum: 0
          maximum: 5000
    PrinterProfileUpdateRequest:
      type: object
      additionalProperties: false
      required:
        [
          expectedVersion,
          name,
          model,
          serviceUuid,
          writeCharacteristicUuid,
          writeMode,
          commandDialect,
          paperWidthMm,
          encoding,
          maxChunkBytes,
          interChunkDelayMs,
          active
        ]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 120
        model:
          type: string
          minLength: 1
          maxLength: 120
        serviceUuid:
          type: string
          maxLength: 64
        writeCharacteristicUuid:
          type: string
          maxLength: 64
        writeMode:
          type: string
          enum: [ WITH_RESPONSE, WITHOUT_RESPONSE ]
        commandDialect:
          type: string
          maxLength: 80
        paperWidthMm:
          type: integer
          enum: [ 58, 80 ]
        encoding:
          type: string
          maxLength: 40
        maxChunkBytes:
          type: integer
          minimum: 1
          maximum: 1024
        interChunkDelayMs:
          type: integer
          minimum: 0
          maximum: 5000
        active:
          type: boolean
        reason:
          type: [ string, 'null' ]
          maxLength: 500
    PrinterProfileResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/PrinterProfile'
    PrinterProfileListResponse:
      type: object
      required: [ data ]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PrinterProfile'
    PrinterPreferenceRequest:
      type: object
      additionalProperties: false
      required: [ printerProfileId ]
      properties:
        printerProfileId:
          $ref: '#/components/schemas/Uuid'
        deviceLabel:
          type: [ string, 'null' ]
          maxLength: 120
        testedBrowser:
          type: [ string, 'null' ]
          maxLength: 120
        testedOs:
          type: [ string, 'null' ]
          maxLength: 120
        lastTestResult:
          type: [ string, 'null' ]
          enum: [ SUCCEEDED, FAILED, UNKNOWN, null ]
    PrinterPreferenceResponse:
      type: object
      required: [ data ]
      properties:
        data:
          type: [ object, 'null' ]
          additionalProperties: false
          required: [ printerProfileId ]
          properties:
            printerProfileId:
              $ref: '#/components/schemas/Uuid'
            deviceLabel:
              type: [ string, 'null' ]
            testedBrowser:
              type: [ string, 'null' ]
            testedOs:
              type: [ string, 'null' ]
            lastTestResult:
              type: [ string, 'null' ]
              enum: [ SUCCEEDED, FAILED, UNKNOWN, null ]
            lastTestedAt:
              type: [ string, 'null' ]
              format: date-time
...
    BusinessSetting:
      type: object
      additionalProperties: false
      required: [ currencyCode, currencyScale, businessTimezone, partnerShareRate, moneyRoundingMode, version ]
      properties:
        currencyCode:
          type: string
          pattern: '^[A-Z]{3}$'
        currencyScale:
          const: 2
        businessTimezone:
          type: string
          maxLength: 100
        partnerShareRate:
          const: '0.500000'
        moneyRoundingMode:
          const: HALF_AWAY_FROM_ZERO
        version:
          type: integer
          minimum: 1
    BusinessSettingUpdateRequest:
      type: object
      additionalProperties: false
      required: [ expectedVersion, currencyCode, businessTimezone, reason ]
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        currencyCode:
          type: string
          pattern: '^[A-Z]{3}$'
        businessTimezone:
          type: string
          minLength: 1
          maxLength: 100
        reason:
          type: string
          minLength: 1
          maxLength: 500
    BusinessSettingResponse:
      type: object
      required: [ data ]
      properties:
        data:
          $ref: '#/components/schemas/BusinessSetting'
...